+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.1.2
 - Fix Misc and Mob Spawn Configs not generating sometimes
+- Use c:chests/wooden Item tag for BWG Chest Boats
 
 # 3.1.1
 - Adjust BWG Blocks to match the same order in the creative tab as vanilla