+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix Misc and Mob Spawn Configs not generating sometimes
 - Use c:chests/wooden Item tag for BWG Chest Boats
 - Fix bonemeal not using up when bonemealing grass blocks on NeoForge
+- Fix incompatibility with Slayer's Beasts, remove duplicate feature from Coniferous Forest(s)
 
 # 3.1.1
 - Adjust BWG Blocks to match the same order in the creative tab as vanilla