+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Use c:chests/wooden Item tag for BWG Chest Boats
 - Fix bonemeal not using up when bonemealing grass blocks on NeoForge
 - Fix incompatibility with Slayer's Beasts, remove duplicate feature from Coniferous Forest(s)
+- Fix Tall Beach and Prairie Grass dropping two tall grass instead of normal grass
 
 # 3.1.1
 - Adjust BWG Blocks to match the same order in the creative tab as vanilla