+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.6.3
 - Fix incompatibility with Slayer's Beasts, remove duplicate feature from Coniferous Forest(s)
 - Fix Tall Beach and Prairie Grass dropping two tall grass instead of normal grass
+- Fix sniffer dropping all bwg loot instead of just one (Forge Only)
 
 # 1.6.2
 - Fix White Puffball not growing over time